separate ReactionsRepository into its interface and implementation
